+++ packages/vendure-plugin-invoices/src/api/invoice-admin.resolver.ts
@@ -36,6 +36,7 @@ export class InvoiceAdminResolver {
     @Ctx() ctx: RequestContext,
     @Args('input') input: InvoiceConfigInput
   ): Promise<InvoiceConfigEntity> {
+    this.invoiceService.throwIfInvalidLicense();
     return this.invoiceService.upsertConfig(ctx, input);
   }
 
@@ -81,6 +82,7 @@ export class InvoiceAdminResolver {
   async invoiceConfig(
     @Ctx() ctx: RequestContext
   ): Promise<InvoiceConfigEntity | undefined> {
+    this.invoiceService.throwIfInvalidLicense();
     return this.invoiceService.getConfig(ctx);
   }
 
@@ -90,6 +92,7 @@ export class InvoiceAdminResolver {
     @Ctx() ctx: RequestContext,
     @Args() args: QueryInvoicesArgs
   ): Promise<PaginatedList<Invoice>> {
+    this.invoiceService.throwIfInvalidLicense();
     return this.invoiceService.findAll(ctx, args.options || undefined);
   }
 }